// web/worker/types.ts
export interface InstallMetrics {
  installCount: number
  installerCount: number
  firstInstallCount: number
  reinstallCount: number
  updateCount: number
  removeCount: number
  failureCount: number
  installs24h: number
  installs7d: number
  installs30d: number
  latestInstallAt: string | null
}